Refactor workflow to workflows for consistency
